+++ i18n/en.ts
@@ -730,19 +730,19 @@ li.checked::marker { content: &quot;\2612&quot;; }
     </message>
     <message>
         <location filename="../resources/ui/mainwindow.ui" line="305"/>
-        <location filename="../resources/ui/mainwindow.ui" line="1259"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1262"/>
         <source>Info</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../resources/ui/mainwindow.ui" line="353"/>
-        <location filename="../resources/ui/mainwindow.ui" line="1289"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1292"/>
         <source>Labels</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../resources/ui/mainwindow.ui" line="329"/>
-        <location filename="../resources/ui/mainwindow.ui" line="1244"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1247"/>
         <source>Files</source>
         <translation type="unfinished"></translation>
     </message>
@@ -977,7 +977,7 @@ li.checked::marker { content: &quot;\2612&quot;; }
     </message>
     <message>
         <location filename="../resources/ui/mainwindow.ui" line="933"/>
-        <location filename="../resources/ui/mainwindow.ui" line="1199"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1202"/>
         <source>Rectangle</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1049,27 +1049,43 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1349"/>
-        <source>Left (Ctrl+B)</source>
+        <location filename="../resources/ui/mainwindow.ui" line="1163"/>
+        <source>Merge (Ctrl+G)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../resources/ui/mainwindow.ui" line="1166"/>
+        <source>Merge multiple shapes (Ctrl+G)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../resources/ui/mainwindow.ui" line="1184"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1187"/>
+        <source>Group selected annotations into one instance (G); split (G)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../resources/ui/mainwindow.ui" line="1352"/>
+        <source>Left (Ctrl+B)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../resources/ui/mainwindow.ui" line="1355"/>
         <source>Ctrl+B</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1370"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1373"/>
         <source>Right (Ctrl+Alt+B)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1373"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1376"/>
         <source>Ctrl+Alt+B</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1394"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1397"/>
         <source>Ctrl+J</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1119,7 +1135,7 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1169"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1190"/>
         <source>G</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1129,63 +1145,63 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1187"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1169"/>
         <source>Ctrl+G</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1274"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1277"/>
         <source>Annotations</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1298"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1301"/>
         <source>Restore</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1310"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1313"/>
         <source>Full screen</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1313"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1316"/>
         <source>Full screen (F11)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1316"/>
         <location filename="../resources/ui/mainwindow.ui" line="1319"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1322"/>
         <source>Enter/Exit full screen mode.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1322"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1325"/>
         <source>F11</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1331"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1334"/>
         <source>KeyPoint</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1346"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1349"/>
         <source>Left</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1367"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1370"/>
         <source>Right</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1388"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1391"/>
         <source>Bottom</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1391"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1394"/>
         <source>Bottom (Ctrl+J)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1195,22 +1211,12 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1163"/>
-        <source>Merge (G)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1166"/>
-        <source>Merge multiple shapes (G)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1219"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1222"/>
         <source>RLE Mask</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1229"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1232"/>
         <source>中文</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1408,12 +1414,7 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1184"/>
-        <source>Group selected annotations into one instance (Ctrl+G); split (Ctrl+G)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../resources/ui/mainwindow.ui" line="1224"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1227"/>
         <source>English</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1446,7 +1447,7 @@ li.checked::marker { content: &quot;\2612&quot;; }
     </message>
     <message>
         <location filename="../resources/ui/mainwindow.ui" line="756"/>
-        <location filename="../resources/ui/mainwindow.ui" line="1211"/>
+        <location filename="../resources/ui/mainwindow.ui" line="1214"/>
         <source>Polygon</source>
         <translation type="unfinished"></translation>
     </message>